+++ base/compiler/tfuncs.jl
@@ -190,7 +190,7 @@ cglobal_tfunc(@nospecialize(fptr)) = Ptr{Cvoid}
 cglobal_tfunc(@nospecialize(fptr), @nospecialize(t)) = (isType(t) ? Ptr{t.parameters[1]} : Ptr)
 cglobal_tfunc(@nospecialize(fptr), t::Const) = (isa(t.val, Type) ? Ptr{t.val} : Ptr)
 add_tfunc(Core.Intrinsics.cglobal, 1, 2, cglobal_tfunc, 5)
-add_tfunc(Core.Intrinsics.select_value, 3, 3,
+add_tfunc(ifelse, 3, 3,
     function (@nospecialize(cnd), @nospecialize(x), @nospecialize(y))
         if isa(cnd, Const)
             if cnd.val === true
@@ -200,29 +200,32 @@ add_tfunc(Core.Intrinsics.select_value, 3, 3,
             else
                 return Bottom
             end
+        elseif isa(cnd, Conditional)
+            # handled in abstract_call
+        elseif !(Bool ⊑ cnd)
+            return Bottom
         end
-        (Bool ⊑ cnd) || return Bottom
         return tmerge(x, y)
     end, 1)
-add_tfunc(===, 2, 2,
-    function (@nospecialize(x), @nospecialize(y))
-        if isa(x, Const) && isa(y, Const)
-            return Const(x.val === y.val)
-        elseif typeintersect(widenconst(x), widenconst(y)) === Bottom
-            return Const(false)
-        elseif (isa(x, Const) && y === typeof(x.val) && isdefined(y, :instance)) ||
-               (isa(y, Const) && x === typeof(y.val) && isdefined(x, :instance))
-            return Const(true)
-        elseif isa(x, Conditional) && isa(y, Const)
-            y.val === false && return Conditional(x.var, x.elsetype, x.vtype)
-            y.val === true && return x
-            return x
-        elseif isa(y, Conditional) && isa(x, Const)
-            x.val === false && return Conditional(y.var, y.elsetype, y.vtype)
-            x.val === true && return y
-        end
-        return Bool
-    end, 1)
+function egal_tfunc(@nospecialize(x), @nospecialize(y))
+    if isa(x, Const) && isa(y, Const)
+        return Const(x.val === y.val)
+    elseif typeintersect(widenconst(x), widenconst(y)) === Bottom
+        return Const(false)
+    elseif (isa(x, Const) && y === typeof(x.val) && isdefined(y, :instance)) ||
+           (isa(y, Const) && x === typeof(y.val) && isdefined(x, :instance))
+        return Const(true)
+    elseif isa(x, Conditional) && isa(y, Const)
+        y.val === false && return Conditional(x.var, x.elsetype, x.vtype)
+        y.val === true && return x
+        return x
+    elseif isa(y, Conditional) && isa(x, Const)
+        x.val === false && return Conditional(y.var, y.elsetype, y.vtype)
+        x.val === true && return y
+    end
+    return Bool
+end
+add_tfunc(===, 2, 2, egal_tfunc, 1)
 function isdefined_tfunc(@nospecialize(args...))
     arg1 = args[1]
     if isa(arg1, Const)
@@ -433,12 +436,46 @@ function const_datatype_getfield_tfunc(sv, fld)
     return nothing
 end
 
+function fieldcount_noerror(@nospecialize t)
+    if t isa UnionAll || t isa Union
+        t = ccall(:jl_argument_datatype, Any, (Any,), t)
+        if t === nothing
+            return nothing
+        end
+        t = t::DataType
+    elseif t == Union{}
+        return 0
+    end
+    if !(t isa DataType)
+        return nothing
+    end
+    if t.name === NamedTuple.body.body.name
+        names, types = t.parameters
+        if names isa Tuple
+            return length(names)
+        end
+        if types isa DataType && types <: Tuple
+            return fieldcount_noerror(types)
+        end
+        abstr = true
+    else
+        abstr = t.abstract || (t.name === Tuple.name && isvatuple(t))
+    end
+    if abstr
+        return nothing
+    end
+    return length(t.types)
+end
+
+
 function try_compute_fieldidx(@nospecialize(typ), @nospecialize(field))
     if isa(field, Symbol)
         field = fieldindex(typ, field, false)
         field == 0 && return nothing
     elseif isa(field, Integer)
-        (1 <= field <= fieldcount(typ)) || return nothing
+        max_fields = fieldcount_noerror(typ)
+        max_fields === nothing && return nothing
+        (1 <= field <= max_fields) || return nothing
     else
         return nothing
     end
@@ -620,8 +657,8 @@ function fieldtype_tfunc(@nospecialize(s0), @nospecialize(name))
     if s0 === Any || s0 === Type || DataType ⊑ s0 || UnionAll ⊑ s0
         return Type
     end
-    # fieldtype only accepts DataType and UnionAll, errors on `Module`
-    if isa(s0,Const) && (!(isa(s0.val,DataType) || isa(s0.val,UnionAll)) || s0.val === Module)
+    # fieldtype only accepts Types, errors on `Module`
+    if isa(s0, Const) && (!(isa(s0.val, DataType) || isa(s0.val, UnionAll) || isa(s0.val, Union)) || s0.val === Module)
         return Bottom
     end
     if s0 == Type{Module} || s0 == Type{Union{}} || isa(s0, Conditional)
@@ -631,12 +668,12 @@ function fieldtype_tfunc(@nospecialize(s0), @nospecialize(name))
     s = instanceof_tfunc(s0)[1]
     u = unwrap_unionall(s)
 
-    if isa(u,Union)
-        return tmerge(rewrap(fieldtype_tfunc(u.a, name),s),
-                      rewrap(fieldtype_tfunc(u.b, name),s))
+    if isa(u, Union)
+        return tmerge(rewrap(fieldtype_tfunc(u.a, name), s),
+                      rewrap(fieldtype_tfunc(u.b, name), s))
     end
 
-    if !isa(u,DataType) || u.abstract
+    if !isa(u, DataType) || u.abstract
         return Type
     end
     if u.name === _NAMEDTUPLE_NAME && !isconcretetype(u)
@@ -656,7 +693,7 @@ function fieldtype_tfunc(@nospecialize(s0), @nospecialize(name))
     end
 
     fld = name.val
-    if isa(fld,Symbol)
+    if isa(fld, Symbol)
         fld = fieldindex(u, fld, false)
     end
     if !isa(fld, Int)
@@ -672,7 +709,7 @@ function fieldtype_tfunc(@nospecialize(s0), @nospecialize(name))
     end
 
     exact = (isa(s0, Const) || isType(s0)) && !has_free_typevars(s)
-    ft = rewrap_unionall(ft,s)
+    ft = rewrap_unionall(ft, s)
     if exact
         return Const(ft)
     end