+++ doc/make.jl
@@ -1,3 +1,9 @@
+if "deploy" in ARGS
+    # Only deploy docs from 64bit Linux to avoid committing multiple versions of the same
+    # docs from different workers.
+    (Sys.ARCH === :x86_64 && Sys.KERNEL === :Linux) || exit()
+end
+
 # Install dependencies needed to build the documentation.
 ENV["JULIA_PKGDIR"] = joinpath(@__DIR__, "deps")
 using Pkg
@@ -51,7 +57,6 @@ const PAGES = [
     "Home" => "index.md",
     hide("NEWS.md"),
     "Manual" => [
-        "manual/introduction.md",
         "manual/getting-started.md",
         "manual/variables.md",
         "manual/integers-and-floating-point-numbers.md",
@@ -154,7 +159,7 @@ makedocs(
     doctest   = ("doctest-fix" in ARGS) ? (:fix) : ("doctest" in ARGS),
     linkcheck = "linkcheck" in ARGS,
     linkcheck_ignore = ["https://bugs.kde.org/show_bug.cgi?id=136779"], # fails to load from nanosoldier?
-    strict    = false,
+    strict    = true,
     checkdocs = :none,
     format    = "pdf" in ARGS ? :latex : :html,
     sitename  = "The Julia Language",
@@ -163,22 +168,18 @@ makedocs(
     pages     = PAGES,
     html_prettyurls = ("deploy" in ARGS),
     html_canonical = ("deploy" in ARGS) ? "https://docs.julialang.org/en/stable/" : nothing,
+    assets = ["assets/julia-manual.css", ]
 )
 
-if "deploy" in ARGS
-    # Only deploy docs from 64bit Linux to avoid committing multiple versions of the same
-    # docs from different workers.
-    (Sys.ARCH === :x86_64 && Sys.KERNEL === :Linux) || return
 
-    # Since the `.travis.yml` config specifies `language: cpp` and not `language: julia` we
-    # need to manually set the version of Julia that we are deploying the docs from.
-    ENV["TRAVIS_JULIA_VERSION"] = "nightly"
+# Since the `.travis.yml` config specifies `language: cpp` and not `language: julia` we
+# need to manually set the version of Julia that we are deploying the docs from.
+ENV["TRAVIS_JULIA_VERSION"] = "nightly"
 
-    deploydocs(
-        repo = "github.com/JuliaLang/julia.git",
-        target = "_build/html/en",
-        dirname = "en",
-        deps = nothing,
-        make = nothing,
-    )
-end
+deploydocs(
+    repo = "github.com/JuliaLang/julia.git",
+    target = "_build/html/en",
+    dirname = "en",
+    deps = nothing,
+    make = nothing,
+)