+++ stdlib/Pkg3/src/Types.jl
@@ -42,10 +42,10 @@ end
 uuid5(namespace::UUID, key::AbstractString) = uuid5(namespace, String(key))
 
 const uuid_dns = UUID(0x6ba7b810_9dad_11d1_80b4_00c04fd430c8)
-const uuid_julia = uuid5(uuid_dns, "julialang.org")
-const uuid_package = uuid5(uuid_julia, "package")
-const uuid_registry = uuid5(uuid_julia, "registry")
-
+const uuid_julia_project = uuid5(uuid_dns, "julialang.org")
+const uuid_package = uuid5(uuid_julia_project, "package")
+const uuid_registry = uuid5(uuid_julia_project, "registry")
+const uuid_julia = uuid5(uuid_package, "julia")
 
 ## user-friendly representation of package IDs ##
 function pkgID(p::UUID, uuid_to_name::Dict{UUID,String})
@@ -384,6 +384,7 @@ end
 
 GitRepo(url::String, revspec) = GitRepo(url, revspec, nothing)
 GitRepo(url::String) = GitRepo(url, "", nothing)
+Base.:(==)(repo1::GitRepo, repo2::GitRepo) = (repo1.url == repo2.url && repo1.rev == repo2.rev && repo1.git_tree_sha1 == repo2.git_tree_sha1)
 
 mutable struct PackageSpec
     name::String
@@ -620,7 +621,7 @@ function read_manifest(io::IO)
     return manifest
 end
 function read_manifest(file::String)
-        try isfile(file) ? open(read_manifest, file) : read_manifest(devnull)
+    try isfile(file) ? open(read_manifest, file) : read_manifest(devnull)
     catch err
         err isa ErrorException && startswith(err.msg, "ambiguious dependency") || rethrow(err)
         err.msg *= "In manifest file: $file"
@@ -1218,16 +1219,16 @@ function manifest_info(env::EnvCache, uuid::UUID)::Union{Dict{String,Any},Nothin
 end
 
 # TODO: redirect to ctx stream
-function printpkgstyle(io::IO, cmd::Symbol, text::String...; ignore_indent=false)
+function printpkgstyle(io::IO, cmd::Symbol, text::String, ignore_indent=false)
     indent = textwidth(string(:Downloaded))
     ignore_indent && (indent = 0)
     printstyled(io, lpad(string(cmd), indent), color=:green, bold=true)
-    println(io, " ", text...)
+    println(io, " ", text)
 end
 
 # TODO: use ctx specific context
-function printpkgstyle(ctx::Context, cmd::Symbol, text::String...; kwargs...)
-    printpkgstyle(stdout, cmd, text...; kwargs...)
+function printpkgstyle(ctx::Context, cmd::Symbol, text::String, ignore_indent=false)
+    printpkgstyle(stdout, cmd, text)
 end